rename function as it was too long
